Fix Go compilation to not depend on SYSROOT location.
When Go is compiled, it embeds the install location (GOROOT) into the
binary, and this location is relative to the build host, so it
contains the sysroot path. This causes problems when using a shared
sstate-cache. If a host pulls the cached package from the
sstate-cache, but has its build dir somewhere else, Go will fail with
a message of the type:

go: cannot find GOROOT directory: /home/user/poky/build/tmp-glibc/sysroots/x86_64-linux/usr/lib/arm-oe-linux-gnueabi/go

We fix this by exporting the right GOROOT environment prior to
building, which Go will use before attempting its built in default.

Note that even with this fix, two hosts with different sysroot
locations will not build identical Go tools. This is because the
location at build time will always be embedded in the binary, but I am
not aware of any way to fix that.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>