classes/go: fix native package dependencies
Building of -native packages is currently broken due to incompatible
naming of go-cross toolchain.

The logic of automatic native DEPENDS, as implemented in classes/native:
expects the toolchain pieces to be named in the following fashion,
<tool>-cross-${TARGET_SYS} (ex. gcc-cross-arm, go-cross-arm). The code
swaps '-cross-' infix to '-native-' in package names listed in DEPENDS.
Go cross package is named 'go-cross', thus there will be no
automatically added dependency.

The patch adds an explicit dependency on go-native for -native packages.

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>