Fix cgo toolchain path error when on target platform.
The path to the gcc and g++ compilers had the yocto build system's
paths rather than a target system path baked into the tools.
This caused broken builds for native code on the target device.
The compile task for the recipe added the package's workspace to
PATH and used explicit paths to the compilers to the batch files
that wrapped access to the tools.