Update golang recipes following go.bbclass changes
caddy doesn't actually build because all these recipes use AUTOREV and
upstream sources have drifted to include new dependencies.  I'm not
interested in maintaining these, but they should continue to serve as
examples of how to use go.bbclass.

(And this also serves as a good example of why we shouldn't use AUTOREV)